Re-organize vm group code
